+++ src/components/Proposal/ProposalData.tsx
@@ -1,7 +1,7 @@
 import { Address, IDAOState, IMemberState, IProposalState, IRewardState, Reward, Stake, Vote } from "@daostack/arc.js";
 import { ethErrorHandler, ethBalance, standardPolling, getArcByDAOAddress } from "lib/util";
 
-import BN = require("bn.js");
+import * as BN from "bn.js";
 import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
 import * as moment from "moment";
 import * as React from "react";