+++ src/components/Redemptions/RedemptionsPage.tsx
@@ -1,7 +1,8 @@
 import { Address, DAOFieldsFragment, IContributionReward, IDAOState, IRewardState, Proposal } from "@daostack/arc.js";
 import { enableWalletProvider, getArcs } from "arc";
 import { redeemProposal } from "actions/arcActions";
-import BN = require("bn.js");
+
+import * as BN from "bn.js";
 import Loading from "components/Shared/Loading";
 import withSubscription, { ISubscriptionProps } from "components/Shared/withSubscription";
 import gql from "graphql-tag";