+++ src/components/Shared/PreTransactionModal.tsx
@@ -1,7 +1,7 @@
 import { IDAOState, IMemberState, IProposalState, IProposalStage } from "@daostack/arc.js";
 import { enableWalletProvider } from "arc";
 
-import BN = require("bn.js");
+import * as BN from "bn.js";
 import classNames from "classnames";
 import Reputation from "components/Account/Reputation";
 import ProposalSummary from "components/Proposal/ProposalSummary";