+++ src/lib/daoHelpers.ts
@@ -1,7 +1,7 @@
 import { DAO, IDAOState, Reputation, Token } from "@daostack/arc.js";
 import { getArc } from "arc";
-import BN = require("bn.js");
 import { Networks } from "./util";
+import * as BN from "bn.js";
 
 // TODO: all this should really be in the client library, after new client is done i will move it there if its not already