+++ CHANGELOG.md
@@ -1,11 +1,22 @@
+## 1.0.3
+  - Features Added
+    - First loading speed improvement
+    - Pages loading process was split to several chunk which is improved UX
+
+  - Bugs Fixed
+    - Integration tests typescript supporting was added
+
 ## 1.0.2
   - Features Added
     - Support Multi-call Generic Scheme
     - Use subgraph v40_0 and arc.js 0.2.76
+    - Added xGEN / GEN bridge
 
   - Bugs Fixed
     - Fix competiotion tab crash
     - Better name for some Competition plugins
+    - Fix wrong link when sharing a proposal
+    - Ignore Network and GraphQL errors up to 10 attempts in a row
 
 ## 1.0.1
   - Bugs Fixed